Remove old-school accelerometer code to replace with iOS shake recognizer Also make "random" story work with the new models -- no more dependence on position in the subreddit array. (Note: Really we should load "random" stories from the reddit API if possible, not just a loaded subreddit)

+++ Classes/Data/SubredditDataModel.h
@@ -17,7 +17,7 @@
     int newsModeIndex;
     
     BOOL canLoadMore;
-    NSInteger totalStories;
+    NSUInteger totalStories;
 }
 @property (nonatomic, readonly) NSString *subreddit;
 @property (nonatomic, readonly) NSMutableArray *stories;
@@ -26,6 +26,6 @@
 
 - (id)initWithSubreddit:(NSString *)subreddit;
 - (NSString *)newsModeString;
-- (int)totalStories;
+- (NSUInteger)totalStories;
 
 @end
